feat: update sitemap to include blog posts and improve URLs
Refactor the sitemap module to include dynamically generated routes for blog posts. This ensures that the sitemap.xml file includes all the blog posts with their respective URLs. Additionally, update the URLs for the blog, projects, and contact pages to use the `URL` constant for consistency and easier maintenance.

Related to recent user commits:
- Add new blog post about top 5 UI libraries with React
- Update image dimensions in blog page
- New section blog and articles (#2)
- Refactor metadata titles in contact and projects pages, and update header component, for better SEO
- Add social media links to the header component
- Add form data for contact
- Refactor metadata and add Google Analytics in layout.tsx
- Write license and readme
- Fix metadata titles in contact and projects pages
- Refactor metadata and about section in layout.tsx and About.tsx files